Legg til periodeoffset på andel
