+++ src/main/resources/db/migration/V65__andel_tilkjent_ytelse_periode_offset.sql
@@ -1,0 +1,1 @@
+ALTER TABLE ANDEL_TILKJENT_YTELSE ADD COLUMN periode_offset BIGINT;
